apps/smoldocling-service/requirements.txt
# Core VLM inference
torch>=2.0.0
transformers>=4.40.0
pillow>=10.0.0
pdf2image>=1.16.0

# Docling integration
docling-core>=1.0.0

# FastAPI framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Observability
prometheus-client>=0.19.0
psutil>=5.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0

# System dependencies note:
# - poppler-utils (pdf2image dependency)
# - libgl1 (OpenCV dependency)
# - CUDA 12.1+ (for GPU inference, optional)
